move methods to the file they are used from
